swap the names as requested
